Add exception handling

+++ src/main/java/ru/kontur/chartographer/service/ChartaService.java
@@ -3,12 +3,13 @@ package ru.kontur.chartographer.service;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
 import ru.kontur.chartographer.domain.Charta;
+import ru.kontur.chartographer.exception.ChartaNotFoundException;
+import ru.kontur.chartographer.exception.InvalidChartaCoordinatesException;
+import ru.kontur.chartographer.exception.RenderImageException;
 import ru.kontur.chartographer.repository.ChartaRepository;
 
 import javax.imageio.ImageIO;
-import java.awt.*;
 import java.awt.image.BufferedImage;
-import java.awt.image.WritableRaster;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
@@ -18,7 +19,6 @@ import java.io.InputStream;
  * @author dzahbarov
  */
 
-
 @Service
 public class ChartaService {
     private final ChartaRepository chartaRepository;
@@ -27,31 +27,35 @@ public class ChartaService {
         this.chartaRepository = chartaRepository;
     }
 
+    public Charta createImage(int width, int height) {
+        return chartaRepository.save(
+                        new Charta(width, height, getBytes(new BufferedImage(width, height, BufferedImage.TYPE_INT_RGB))));
+    }
 
-    public byte[] getPic(String id, int x, int y, int width, int height) {
-        Long id_long = Long.parseLong(id);
-        Charta chartaFromDb = chartaRepository.findById(id_long).orElse(null);
-        InputStream is = new ByteArrayInputStream(chartaFromDb.getImage());
-        try {
-            BufferedImage imageFromDb = ImageIO.read(is);
-            BufferedImage subimage = imageFromDb.getSubimage(x, y, width, height);
-            return getBytes(subimage);
+    public byte[] getSubCharta(long id, int x, int y, int width, int height) {
+        Charta chartaFromDb = chartaRepository.findById(id)
+                .orElseThrow(() -> new ChartaNotFoundException("Charta with id " + id + " is not found"));
+
+        validateCoordinates(chartaFromDb, x, y, width, height);
+
+        // может выезжать за границы
+
+        try (InputStream is = new ByteArrayInputStream(chartaFromDb.getImage())) {
+            return getBytes(ImageIO.read(is).getSubimage(x, y, width, height));
         } catch (IOException e) {
-            e.printStackTrace();
+            throw new RenderImageException("Can't render image");
         }
-        return new byte[]{};
     }
 
-    public Long save(BufferedImage image) {
-        byte[] bytes = getBytes(image);
-        Charta charta = new Charta();
-        charta.setImage(bytes);
-        Charta chartaFromDb = chartaRepository.save(charta);
-        return chartaFromDb.getId();
-    }
 
-    public void updateCharta(String id, int x, int y, int width, int height, MultipartFile image) {
-        Charta chartaFromDb = chartaRepository.findById(Long.parseLong(id)).orElse(null);
+    public void updateCharta(long id, int x, int y, int width, int height, MultipartFile image) {
+
+        // Тут плохо
+        Charta chartaFromDb = chartaRepository.findById(id)
+                .orElseThrow(() -> new ChartaNotFoundException("Charta with id " + id + " is not found"));
+
+        validateCoordinates(chartaFromDb, x, y, width, height);
+
         InputStream is = new ByteArrayInputStream(chartaFromDb.getImage());
 //        BigBufferedImage image = new BigBufferedImage()
         try {
@@ -61,8 +65,8 @@ public class ChartaService {
             BufferedImage newImage = ImageIO.read(newImageIs);
 
             for (int i = y; i < x + height; i++) {
-                for (int j = x; j < y + width; j++) {;
-                    imageFromDb.setRGB(i, j,  newImage.getRGB(i-y, j-x));
+                for (int j = x; j < y + width; j++) {
+                    imageFromDb.setRGB(i, j, newImage.getRGB(i - y, j - x));
                 }
             }
 
@@ -74,30 +78,39 @@ public class ChartaService {
     }
 
     public void deleteImage(long id) {
+        if (!chartaRepository.existsById(id)) {
+            throw new ChartaNotFoundException("Charta with id " + id + " is not found");
+        }
         chartaRepository.deleteById(id);
     }
 
     private byte[] getBytes(BufferedImage image) {
-        ByteArrayOutputStream baos = new ByteArrayOutputStream();
+        ByteArrayOutputStream byteStream = new ByteArrayOutputStream();
         try {
-            ImageIO.write(image, "bmp", baos);
+            ImageIO.write(image, "bmp", byteStream);
         } catch (IOException e) {
-            e.printStackTrace();
+            throw new RenderImageException("Can't render image");
         }
-        return baos.toByteArray();
+        return byteStream.toByteArray();
     }
 
-    public String save(MultipartFile image) {
-        Charta charta = new Charta();
-        try {
-            charta.setImage(image.getBytes());
-        } catch (IOException e) {
-            e.printStackTrace();
+    private void validateCoordinates(Charta chartaFromDb, int x, int y, int width, int height) {
+        if (x > chartaFromDb.getWidth() || y > chartaFromDb.getHeight() || x + width < 0 || y + height < 0) {
+            throw new InvalidChartaCoordinatesException("Invalid coordinates for updating");
         }
-        return Long.toString(chartaRepository.save(charta).getId());
     }
 
-    public byte[] getPic(String id) {
-        return chartaRepository.findById(Long.parseLong(id)).orElse(null).getImage();
-    }
+//    public String save(MultipartFile image) {
+//        Charta charta = new Charta();
+//        try {
+//            charta.setImage(image.getBytes());
+//        } catch (IOException e) {
+//            e.printStackTrace();
+//        }
+//        return Long.toString(chartaRepository.save(charta).getId());
+//    }
+
+//    public byte[] getPic(long id) {
+//        return chartaRepository.findById(id).orElse(null).getImage();
+//    }
 }
